requirements.txt
# Core Data Science Libraries
pandas>=2.0.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.12.0

# Machine Learning Libraries
scikit-learn>=1.3.0
catboost>=1.2.0

# Data Processing
openpyxl>=3.1.0
xlrd>=2.0.0
python-dateutil>=2.8.0

# Model Persistence and Utilities
joblib>=1.3.0

# Web API Framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
python-multipart>=0.0.5
aiofiles>=0.8.0

# HTTP Client for API Integration
requests>=2.28.0
aiohttp>=3.8.0

# Async Support
asyncio-mqtt>=0.11.0

# Configuration Management
pyyaml>=6.0

# Jupyter Notebook Support
jupyter>=1.0.0
jupyterlab>=3.4.0
ipykernel>=6.15.0
notebook>=6.4.0

# Development and Testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0

# Logging and Utilities
python-json-logger>=2.0.0
typing-extensions>=4.3.0

# Enhanced Data Visualization
plotly>=5.10.0
